services/api/hackathon24Team1/requirements.txt
Django==4.2.16
django-cors-headers==4.4.0
django-import-export==4.1.1
djangorestframework==3.15.2
